Complete initial map.mcfunction migration
This is the result of my initial migration from Command Center 3 (CC3) to an mcfunction file. The idea is to remove the whole cumbersome concept of command centers into functions and this commit represents the full conversion of my latest and largest command center into a single file.

There are issues here. The first book already doesn't work. But I wanted to commit this initial migration as a milestone.