Add setblock/air commands to each block
The issue with (nearly?) all setblock commands with replace option
is that they didn't work. Like, for a chest, sometimes the chest
would be created but it would be empty. Or, for books on lecterns,
sometimes it would work (if the lectern was nearby) and other times
the book itself would be un-openable.

Solved this with the hack of setting the block(s) to "air" first
before creating the thing I actually need.

I'm sure there's a better way to do this, but this brute force
method is working just fine so far.